+++ com.unity.render-pipelines.core/Runtime/Lighting/ProbeVolume/ProbeVolumeSceneData.cs
@@ -9,9 +9,8 @@ using UnityEditor;
 namespace UnityEngine.Experimental.Rendering
 {
     // Add Profile and baking settings.
-
-    [System.Serializable]
     /// <summary> A class containing info about the bounds defined by the probe volumes in various scenes. </summary>
+    [System.Serializable]
     public class ProbeVolumeSceneData : ISerializationCallbackReceiver
     {
         static PropertyInfo s_SceneGUID = typeof(Scene).GetProperty("guid", System.Reflection.BindingFlags.NonPublic | BindingFlags.Instance);
@@ -91,8 +90,10 @@ namespace UnityEngine.Experimental.Rendering
             }
         }
 
-        /// <summary>Constructor for ProbeVolumeSceneData. </summary>
-        /// <param name="parentAsset">The asset holding this ProbeVolumeSceneData, it will be dirtied every time scene bounds or settings are changed. </param>
+        /// <summary>
+        /// Constructor for ProbeVolumeSceneData.
+        /// </summary>
+        /// <param name="parentAsset">The asset holding this ProbeVolumeSceneData, it will be dirtied every time scene bounds or settings are changed.</param>
         /// <param name="parentSceneDataPropertyName">The name of the property holding the ProbeVolumeSceneData in the parentAsset.</param>
         public ProbeVolumeSceneData(Object parentAsset, string parentSceneDataPropertyName)
         {
@@ -114,7 +115,7 @@ namespace UnityEngine.Experimental.Rendering
         }
 
         /// <summary>Set a reference to the object holding this ProbeVolumeSceneData.</summary>
-        /// <param name="parentAsset">The object holding this ProbeVolumeSceneData, it will be dirtied every time scene bounds or settings are changed. </param>
+        /// <param name="parent">The object holding this ProbeVolumeSceneData, it will be dirtied every time scene bounds or settings are changed. </param>
         /// <param name="parentSceneDataPropertyName">The name of the property holding the ProbeVolumeSceneData in the parentAsset.</param>
         public void SetParentObject(Object parent, string parentSceneDataPropertyName)
         {