+++ com.unity.render-pipelines.core/CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [12.1.3] - 2021-11-17
 
-Version Updated
-The version number for this package has increased due to a version update of a related graphics package.
+### Fixed
+- Fixed XR support in CoreUtils.DrawFullscreen function.
 
 ## [12.1.2] - 2021-10-22